requirements.txt
# Dependecies for web scraping
bs4
requests
# Telethon Dependecies
Telethon==1.10.10
cryptg
pillow
# YAML
PyYAML
# aiohttp
aiohttp[speedups]
# sentry logging
sentry-sdk==0.14.1
# discord bot
discord.py